--- work/org/apache/jsp/WEB_002dINF/view/billboard/show_jsp.java
@@ -1,219 +1,0 @@
-package org.apache.jsp.WEB_002dINF.view.billboard;
-
-import javax.servlet.*;
-import javax.servlet.http.*;
-import javax.servlet.jsp.*;
-
-public final class show_jsp extends org.apache.jasper.runtime.HttpJspBase
-    implements org.apache.jasper.runtime.JspSourceDependent {
-
-static private org.apache.jasper.runtime.ProtectedFunctionMapper _jspx_fnmap_0;
-static private org.apache.jasper.runtime.ProtectedFunctionMapper _jspx_fnmap_1;
-static private org.apache.jasper.runtime.ProtectedFunctionMapper _jspx_fnmap_2;
-
-static {
-  _jspx_fnmap_0= org.apache.jasper.runtime.ProtectedFunctionMapper.getMapForFunction("f:url", org.seasar.struts.taglib.S2Functions.class, "url", new Class[] {java.lang.String.class});
-  _jspx_fnmap_1= org.apache.jasper.runtime.ProtectedFunctionMapper.getMapForFunction("f:h", org.seasar.struts.taglib.S2Functions.class, "h", new Class[] {java.lang.Object.class});
-  _jspx_fnmap_2= org.apache.jasper.runtime.ProtectedFunctionMapper.getInstance();
-  _jspx_fnmap_2.mapFunction("f:br", org.seasar.struts.taglib.S2Functions.class, "br", new Class[] {java.lang.String.class});
-  _jspx_fnmap_2.mapFunction("f:h", org.seasar.struts.taglib.S2Functions.class, "h", new Class[] {java.lang.Object.class});
-}
-
-  private static final JspFactory _jspxFactory = JspFactory.getDefaultFactory();
-
-  private static java.util.List _jspx_dependants;
-
-  static {
-    _jspx_dependants = new java.util.ArrayList(1);
-    _jspx_dependants.add("/WEB-INF/view/common/common.jsp");
-  }
-
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fhtml_005ferrors_005fnobody;
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fs_005fform;
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fhtml_005fhidden_0026_005fproperty_005fnobody;
-
-  private javax.el.ExpressionFactory _el_expressionfactory;
-  private org.apache.AnnotationProcessor _jsp_annotationprocessor;
-
-  public Object getDependants() {
-    return _jspx_dependants;
-  }
-
-  public void _jspInit() {
-    _005fjspx_005ftagPool_005fhtml_005ferrors_005fnobody = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _005fjspx_005ftagPool_005fs_005fform = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _005fjspx_005ftagPool_005fhtml_005fhidden_0026_005fproperty_005fnobody = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _el_expressionfactory = _jspxFactory.getJspApplicationContext(getServletConfig().getServletContext()).getExpressionFactory();
-    _jsp_annotationprocessor = (org.apache.AnnotationProcessor) getServletConfig().getServletContext().getAttribute(org.apache.AnnotationProcessor.class.getName());
-  }
-
-  public void _jspDestroy() {
-    _005fjspx_005ftagPool_005fhtml_005ferrors_005fnobody.release();
-    _005fjspx_005ftagPool_005fs_005fform.release();
-    _005fjspx_005ftagPool_005fhtml_005fhidden_0026_005fproperty_005fnobody.release();
-  }
-
-  public void _jspService(HttpServletRequest request, HttpServletResponse response)
-        throws java.io.IOException, ServletException {
-
-    PageContext pageContext = null;
-    HttpSession session = null;
-    ServletContext application = null;
-    ServletConfig config = null;
-    JspWriter out = null;
-    Object page = this;
-    JspWriter _jspx_out = null;
-    PageContext _jspx_page_context = null;
-
-
-    try {
-      response.setContentType("text/html;charset=UTF-8");
-      pageContext = _jspxFactory.getPageContext(this, request, response,
-      			null, true, 8192, true);
-      _jspx_page_context = pageContext;
-      application = pageContext.getServletContext();
-      config = pageContext.getServletConfig();
-      session = pageContext.getSession();
-      out = pageContext.getOut();
-      _jspx_out = out;
-
-      out.write("\r\n");
-      out.write("\r\n");
-      out.write("\r\n");
-      out.write("\r\n");
-      out.write("  \r\n");
-      out.write("  \r\n");
-      out.write("\r\n");
-      out.write("\r\n");
-      out.write("<!DOCTYPE html PUBLIC \"-//W3C//DTD XHTML 1.0 Transitional//EN\" \"http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd\">\r\n");
-      out.write("<html>\r\n");
-      out.write("<head>\r\n");
-      out.write("<meta http-equiv=\"Content-Type\" content=\"text/html; charset=UTF-8\" />\r\n");
-      out.write("<title>詳細画面</title>\r\n");
-      out.write("<link rel=\"stylesheet\" type=\"text/css\" href=\"");
-      out.write((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${f:url('/css/style.css')}", java.lang.String.class, (PageContext)_jspx_page_context, _jspx_fnmap_0, false));
-      out.write("\" />\r\n");
-      out.write("</head>\r\n");
-      out.write("\r\n");
-      out.write("<body>\r\n");
-      out.write("\r\n");
-      out.write("\t");
-      if (_jspx_meth_html_005ferrors_005f0(_jspx_page_context))
-        return;
-      out.write('\r');
-      out.write('\n');
-      out.write('	');
-      if (_jspx_meth_s_005fform_005f0(_jspx_page_context))
-        return;
-      out.write("\r\n");
-      out.write("</body>\r\n");
-      out.write("\r\n");
-      out.write("</html>");
-    } catch (Throwable t) {
-      if (!(t instanceof SkipPageException)){
-        out = _jspx_out;
-        if (out != null && out.getBufferSize() != 0)
-          try { out.clearBuffer(); } catch (java.io.IOException e) {}
-        if (_jspx_page_context != null) _jspx_page_context.handlePageException(t);
-        else log(t.getMessage(), t);
-      }
-    } finally {
-      _jspxFactory.releasePageContext(_jspx_page_context);
-    }
-  }
-
-  private boolean _jspx_meth_html_005ferrors_005f0(PageContext _jspx_page_context)
-          throws Throwable {
-    PageContext pageContext = _jspx_page_context;
-    JspWriter out = _jspx_page_context.getOut();
-    //  html:errors
-    org.apache.struts.taglib.html.ErrorsTag _jspx_th_html_005ferrors_005f0 = (org.apache.struts.taglib.html.ErrorsTag) _005fjspx_005ftagPool_005fhtml_005ferrors_005fnobody.get(org.apache.struts.taglib.html.ErrorsTag.class);
-    _jspx_th_html_005ferrors_005f0.setPageContext(_jspx_page_context);
-    _jspx_th_html_005ferrors_005f0.setParent(null);
-    int _jspx_eval_html_005ferrors_005f0 = _jspx_th_html_005ferrors_005f0.doStartTag();
-    if (_jspx_th_html_005ferrors_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-      _005fjspx_005ftagPool_005fhtml_005ferrors_005fnobody.reuse(_jspx_th_html_005ferrors_005f0);
-      return true;
-    }
-    _005fjspx_005ftagPool_005fhtml_005ferrors_005fnobody.reuse(_jspx_th_html_005ferrors_005f0);
-    return false;
-  }
-
-  private boolean _jspx_meth_s_005fform_005f0(PageContext _jspx_page_context)
-          throws Throwable {
-    PageContext pageContext = _jspx_page_context;
-    JspWriter out = _jspx_page_context.getOut();
-    //  s:form
-    org.seasar.struts.taglib.S2FormTag _jspx_th_s_005fform_005f0 = (org.seasar.struts.taglib.S2FormTag) _005fjspx_005ftagPool_005fs_005fform.get(org.seasar.struts.taglib.S2FormTag.class);
-    _jspx_th_s_005fform_005f0.setPageContext(_jspx_page_context);
-    _jspx_th_s_005fform_005f0.setParent(null);
-    int _jspx_eval_s_005fform_005f0 = _jspx_th_s_005fform_005f0.doStartTag();
-    if (_jspx_eval_s_005fform_005f0 != javax.servlet.jsp.tagext.Tag.SKIP_BODY) {
-      do {
-        out.write("\r\n");
-        out.write("\t\t");
-        if (_jspx_meth_html_005fhidden_005f0(_jspx_th_s_005fform_005f0, _jspx_page_context))
-          return true;
-        out.write("\r\n");
-        out.write("\t\t<div id=\"main\" class=\"container\">\r\n");
-        out.write("\t\t\t<h3 align=\"center\">掲示板</h3>\r\n");
-        out.write("\t\t\t<hr size=\"4\" />\r\n");
-        out.write("\t\t\t<div align=\"right\">\r\n");
-        out.write("\t\t\t\t<input type=\"submit\" name=\"index\" value=\"戻る\" />\r\n");
-        out.write("\t\t\t</div>\r\n");
-        out.write("\t\t\t<div class=\"mid_title\">タイトル</div>\r\n");
-        out.write("\t\t\t<div class=\"report\">");
-        out.write((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${f:h(billboardItem.title)}", java.lang.String.class, (PageContext)_jspx_page_context, _jspx_fnmap_1, false));
-        out.write("</div>\r\n");
-        out.write("\t\t\t<div class=\"mid_title\">内容</div>\r\n");
-        out.write("\t\t\t<div class=\"report\">");
-        out.write((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${f:br(f:h(billboardItem.note))}", java.lang.String.class, (PageContext)_jspx_page_context, _jspx_fnmap_2, false));
-        out.write("</div>\r\n");
-        out.write("\t\t\t<div class=\"mid_title\">分類</div>\r\n");
-        out.write("\t\t\t<div class=\"report\">");
-        out.write((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${f:h(billboardItem.nameMst.name)}", java.lang.String.class, (PageContext)_jspx_page_context, _jspx_fnmap_2, false));
-        out.write("</div>\r\n");
-        out.write("\t\t\t<div class=\"mid_title\">作成者</div>\r\n");
-        out.write("\t\t\t<div class=\"report\">");
-        out.write((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${f:h(billboardItem.user.name)}", java.lang.String.class, (PageContext)_jspx_page_context, _jspx_fnmap_2, false));
-        out.write("</div>\r\n");
-        out.write("\r\n");
-        out.write("\t\t\t<hr size=\"1\" />\r\n");
-        out.write("\t\t\t<div align=\"center\">\r\n");
-        out.write("\t\t\t\t<input type=\"submit\" name=\"edit\" value=\"編集\" /> <input type=\"submit\"\r\n");
-        out.write("\t\t\t\t\tname=\"index\" value=\"キャンセル\" />\r\n");
-        out.write("\t\t\t</div>\r\n");
-        out.write("\t\t</div>\r\n");
-        out.write("\t");
-        int evalDoAfterBody = _jspx_th_s_005fform_005f0.doAfterBody();
-        if (evalDoAfterBody != javax.servlet.jsp.tagext.BodyTag.EVAL_BODY_AGAIN)
-          break;
-      } while (true);
-    }
-    if (_jspx_th_s_005fform_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-      _005fjspx_005ftagPool_005fs_005fform.reuse(_jspx_th_s_005fform_005f0);
-      return true;
-    }
-    _005fjspx_005ftagPool_005fs_005fform.reuse(_jspx_th_s_005fform_005f0);
-    return false;
-  }
-
-  private boolean _jspx_meth_html_005fhidden_005f0(javax.servlet.jsp.tagext.JspTag _jspx_th_s_005fform_005f0, PageContext _jspx_page_context)
-          throws Throwable {
-    PageContext pageContext = _jspx_page_context;
-    JspWriter out = _jspx_page_context.getOut();
-    //  html:hidden
-    org.apache.struts.taglib.html.HiddenTag _jspx_th_html_005fhidden_005f0 = (org.apache.struts.taglib.html.HiddenTag) _005fjspx_005ftagPool_005fhtml_005fhidden_0026_005fproperty_005fnobody.get(org.apache.struts.taglib.html.HiddenTag.class);
-    _jspx_th_html_005fhidden_005f0.setPageContext(_jspx_page_context);
-    _jspx_th_html_005fhidden_005f0.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_s_005fform_005f0);
-    // /WEB-INF/view/billboard/show.jsp(13,2) name = property type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_html_005fhidden_005f0.setProperty("id");
-    int _jspx_eval_html_005fhidden_005f0 = _jspx_th_html_005fhidden_005f0.doStartTag();
-    if (_jspx_th_html_005fhidden_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-      _005fjspx_005ftagPool_005fhtml_005fhidden_0026_005fproperty_005fnobody.reuse(_jspx_th_html_005fhidden_005f0);
-      return true;
-    }
-    _005fjspx_005ftagPool_005fhtml_005fhidden_0026_005fproperty_005fnobody.reuse(_jspx_th_html_005fhidden_005f0);
-    return false;
-  }
-}
